refactor: split rest_store into submodules; address review feedback
- Split rest_store.rs into submodules (auth, client, contract, store, tests)
  for readability.
- Drop the RestFileApiContract trait; its methods now live on RestEndpointConfig
  (the sole contract), removing speculative generality.
- The store owns 404 -> NotFound (via status_error); RestEndpointConfig::map_status
  maps only non-404 codes (e.g. 409 -> AlreadyExists), so 404 handling no longer
  depends on the contract config getting it right.
- Rename HEADER_REFRESH_SKEW -> HEADER_REFRESH_BUFFER and clarify its doc.
